Use /etc/hosts to resolve to machine ip on machine
The DNS server only receives requests from the WiFi clients.

Making the flipcam machine use that DNS server would be potentially
mess with the existing network configuration. It also would be very
difficult to implement it in an agnostic way.

What was considered is making dnsmasq forward the requests to the
nameservers that were configured when flipcam starts but this runs
into issues with fighting for /etc/resolv.conf and a potential need
for openresolv. Changing /etc/hosts is way simpler though changes
are only made during the installation phase. This means that the
domain configured with --hostname will keep resolving to the ip
even though no interface is configured when flipcam is not running.
The downside of this seems neglible.